使用策略模式将业务与具体的MQ操作分离
