+++ xxpay4spring-boot/src/main/java/org/xxpay/boot/service/mq/MqConfig.java
@@ -3,6 +3,7 @@ package org.xxpay.boot.service.mq;
 import org.apache.activemq.command.ActiveMQQueue;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 import javax.jms.Queue;
 
@@ -17,10 +18,10 @@ import javax.jms.Queue;
 public class MqConfig {
 
     public static final String PAY_NOTIFY_QUEUE_NAME = "pay.notify.queue";
-
-    @Bean
-    public Queue payNotifyQueue() {
-        return new ActiveMQQueue(PAY_NOTIFY_QUEUE_NAME);
+    
+    public static class Impl{
+    	public static final String ACTIVE_MQ = "activeMQ";
+    	public static final String RABBIT_MQ = "rabbitMQ";
     }
-
+    
 }
